+++ src/admin/FinAccount/finAccountService.js
@@ -10,8 +10,8 @@ import "dotenv/config";
 import { logger } from "../../../config/winston";
 
 export const createFinAccount = async (adminIdx, finAccountCategoryIdx, isProfit, finAccountItem, finAccountCost, finAccountDate, etc) => {
+  const connection = await pool.getConnection(async (conn) => conn);
   try {
-    const connection = await pool.getConnection(async (conn) => conn);
     const finAccountInfo = [adminIdx, finAccountCategoryIdx, isProfit, finAccountItem, finAccountCost, finAccountDate, etc];
 
     const userInfoRows = await authProvider.statusCheckByIdx(adminIdx);
@@ -23,24 +23,30 @@ export const createFinAccount = async (adminIdx, finAccountCategoryIdx, isProfit
 
     //todo: finAcountCategoryIdx 상태 확인
     const categoryInfoRows = await accountProvider.categoryStatusCheck(finAccountCategoryIdx);
-    console.log(finAccountCategoryIdx);
+    if (!categoryInfoRows[0]) {
+      return errResponse(baseResponse.FINACCOUNT_CATEGORY_NOT_EXIST);
+    }
+    if (categoryInfoRows[0].adminIdx !== parseInt(adminIdx)) {
+      return errResponse(baseResponse.FINACCOUNT_CATEGORY_NOT_IN_CLUB);
+    }
     if (categoryInfoRows[0].status === "INACTIVE") {
       return errResponse(baseResponse.SIGNIN_INACTIVE_ACCOUNT);
     } else if (categoryInfoRows[0].status === "DELETED") {
       return errResponse(baseResponse.FINACCOUNT_CATEGORY_DELETED);
     }
     const createAccountResult = await accountDao.insertFinAccount(connection, finAccountInfo);
-    connection.release();
     return response(baseResponse.SUCCESS, createAccountResult[0].insertId);
   } catch (err) {
     logger.error(`Admin - createFinAccount Service error: ${err.message}`);
     return errResponse(baseResponse.DB_ERROR);
+  } finally {
+    connection.release();
   }
 };
 
 export const createFinAccCategory = async (categoryName, adminIdx) => {
+  const connection = await pool.getConnection(async (conn) => conn);
   try {
-    const connection = await pool.getConnection(async (conn) => conn);
     const finAccCategoryInfo = [categoryName, adminIdx];
     //todo: admin 상태 확인
     const userInfoRows = await authProvider.statusCheckByIdx(adminIdx);
@@ -53,20 +59,36 @@ export const createFinAccCategory = async (categoryName, adminIdx) => {
     const categoryRows = await accountProvider.categoryDupCheck(adminIdx, categoryName);
     if (categoryRows.length !== 0) return errResponse(baseResponse.FINACCOUNT_CATEGORY_EXIST);
     const createAccCategoryResult = await accountDao.insertFinAccCategory(connection, finAccCategoryInfo);
-    connection.release();
     return response(baseResponse.SUCCESS, createAccCategoryResult[0].insertId);
   } catch (err) {
     logger.error(`Admin - createFinAccCategory Service error: ${err.message}`);
     return errResponse(baseResponse.DB_ERROR);
+  } finally {
+    connection.release();
   }
 };
 
 export const updateFinCategory = async (adminIdx, categroyIdx, categoryName) => {
+  const connection = await pool.getConnection(async (conn) => conn);
   try {
-    const connection = await pool.getConnection(async (conn) => conn);
     const finAccCategoryInfo = [categoryName, categroyIdx];
     //todo: admin 상태 확인
     const userInfoRows = await authProvider.statusCheckByIdx(adminIdx);
+
+    const categoryInfoRows = await accountProvider.categoryStatusCheck(categroyIdx);
+    if (!categoryInfoRows[0]) {
+      return errResponse(baseResponse.FINACCOUNT_CATEGORY_NOT_EXIST);
+    }
+
+    if (categoryInfoRows[0].adminIdx !== parseInt(adminIdx)) {
+      return errResponse(baseResponse.FINACCOUNT_CATEGORY_NOT_IN_CLUB);
+    }
+    if (categoryInfoRows[0].status === "INACTIVE") {
+      return errResponse(baseResponse.SIGNIN_INACTIVE_ACCOUNT);
+    } else if (categoryInfoRows[0].status === "DELETED") {
+      return errResponse(baseResponse.FINACCOUNT_CATEGORY_DELETED);
+    }
+
     if (userInfoRows[0].status === "INACTIVE") {
       return errResponse(baseResponse.SIGNIN_INACTIVE_ACCOUNT);
     } else if (userInfoRows[0].status === "DELETED") {
@@ -78,17 +100,18 @@ export const updateFinCategory = async (adminIdx, categroyIdx, categoryName) => 
       return errResponse(baseResponse.FINACCOUNT_CATEGORY_EXIST);
     }
     const updateAccCategoryResult = await accountDao.modifyFinAccCategory(connection, finAccCategoryInfo);
-    connection.release();
     return response(baseResponse.SUCCESS, updateAccCategoryResult[0].insertId);
   } catch (err) {
     logger.error(`Admin - updateFinCategory Service error: ${err.message}`);
     return errResponse(baseResponse.DB_ERROR);
+  } finally {
+    connection.release();
   }
 };
 
 export const updateFinAccount = async (accountIdx, adminIdx, finAccountCategoryIdx, finAccountItem, isProfit, finAccountCost, finAccountDate, etc) => {
+  const connection = await pool.getConnection(async (conn) => conn);
   try {
-    const connection = await pool.getConnection(async (conn) => conn);
     const finAccountInfo = [finAccountCategoryIdx, finAccountItem, isProfit, finAccountCost, finAccountDate, etc, accountIdx];
     //todo: adminIdx 상태 확인
     const userInfoRows = await authProvider.statusCheckByIdx(adminIdx);
@@ -99,27 +122,37 @@ export const updateFinAccount = async (accountIdx, adminIdx, finAccountCategoryI
     }
     const finAccountInfoRows = await accountProvider.accountStatusCheck(accountIdx);
     if (finAccountInfoRows.length === 0) return errResponse(baseResponse.FINACCOUNT_NOT_EXIST);
+    if (finAccountInfoRows[0].adminIdx !== parseInt(adminIdx)) return errResponse(baseResponse.FINACCOUNT_NOT_IN_CLUB);
     if (finAccountInfoRows[0].status === "DELETED") return errResponse(baseResponse.FINACCOUNT_ALREADY_DELETED);
 
     //todo: 카테고리 idx 존재하는 지 확인
     const categoryInfoRows = await accountProvider.categoryStatusCheck(finAccountCategoryIdx);
+
+    if (!categoryInfoRows[0]) {
+      return errResponse(baseResponse.FINACCOUNT_CATEGORY_NOT_EXIST);
+    }
+
+    if (categoryInfoRows[0].adminIdx !== parseInt(adminIdx)) {
+      return errResponse(baseResponse.FINACCOUNT_CATEGORY_NOT_IN_CLUB);
+    }
     if (categoryInfoRows[0].status === "INACTIVE") {
       return errResponse(baseResponse.SIGNIN_INACTIVE_ACCOUNT);
     } else if (categoryInfoRows[0].status === "DELETED") {
       return errResponse(baseResponse.FINACCOUNT_CATEGORY_DELETED);
     }
     const updateFinAccountResult = await accountDao.modifyFinAccount(connection, finAccountInfo);
-    connection.release();
     return response(baseResponse.SUCCESS, updateFinAccountResult[0].insertId);
   } catch (err) {
     logger.error(`Admin - updateFinAccount Service error: ${err.message}`);
     return errResponse(baseResponse.DB_ERROR);
+  } finally {
+    connection.release();
   }
 };
 
 export const deleteFinAccount = async (accountIdx, adminIdx) => {
+  const connection = await pool.getConnection(async (conn) => conn);
   try {
-    const connection = await pool.getConnection(async (conn) => conn);
     const finAccountInfo = [accountIdx];
     //todo: adminIdx 상태 확인
     const userInfoRows = await authProvider.statusCheckByIdx(adminIdx);
@@ -132,13 +165,43 @@ export const deleteFinAccount = async (accountIdx, adminIdx) => {
     //todo: accountIdx 존재하는 지 확인
     const finAccountInfoRows = await accountProvider.accountStatusCheck(accountIdx);
     if (finAccountInfoRows.length === 0) return errResponse(baseResponse.FINACCOUNT_NOT_EXIST);
+    if (finAccountInfoRows[0].adminIdx !== parseInt(adminIdx)) return errResponse(baseResponse.FINACCOUNT_NOT_IN_CLUB);
     if (finAccountInfoRows[0].status === "DELETED") return errResponse(baseResponse.FINACCOUNT_ALREADY_DELETED);
 
     const deleteFinAccountResult = await accountDao.deleteFinAccount(connection, finAccountInfo);
-    connection.release();
     return response(baseResponse.SUCCESS, deleteFinAccountResult[0].insertId);
   } catch (err) {
     logger.error(`Admin - deleteFinAccount Service error: ${err.message}`);
     return errResponse(baseResponse.DB_ERROR);
+  } finally {
+    connection.release();
+  }
+};
+
+export const getFinAccountByIdx = async (accountIdx, adminIdx) => {
+  const connection = await pool.getConnection(async (conn) => conn);
+  try {
+    const finAccountInfo = [accountIdx];
+    //todo: adminIdx 상태 확인
+    const userInfoRows = await authProvider.statusCheckByIdx(adminIdx);
+    if (userInfoRows[0].status === "INACTIVE") {
+      return errResponse(baseResponse.SIGNIN_INACTIVE_ACCOUNT);
+    } else if (userInfoRows[0].status === "DELETED") {
+      return errResponse(baseResponse.SIGNIN_WITHDRAWAL_ACCOUNT);
+    }
+
+    //todo: accountIdx 존재하는 지 확인
+    const finAccountInfoRows = await accountProvider.accountStatusCheck(accountIdx);
+    if (finAccountInfoRows.length === 0) return errResponse(baseResponse.FINACCOUNT_NOT_EXIST);
+    if (finAccountInfoRows[0].adminIdx !== parseInt(adminIdx)) return errResponse(baseResponse.FINACCOUNT_NOT_IN_CLUB);
+    if (finAccountInfoRows[0].status === "DELETED") return errResponse(baseResponse.FINACCOUNT_ALREADY_DELETED);
+
+    const getFinAccountResult = await accountDao.getFinAccountByIdx(connection, finAccountInfo);
+    return response(baseResponse.SUCCESS, getFinAccountResult[0]);
+  } catch (err) {
+    logger.error(`Admin - deleteFinAccount Service error: ${err.message}`);
+    return errResponse(baseResponse.DB_ERROR);
+  } finally {
+    connection.release();
   }
 };
